arduino_bridge/pyproject.toml
[project]
name = "arduino-bridge"
version = "1.0.0"
description = "Desktop-Tool für Arduino-Firmware-Flash"
requires-python = ">=3.11"
dependencies = [
    "PyQt6>=6.0.0",
    "pyserial>=3.0",
    "websockets>=12.0",
]

[project.scripts]
arduino-bridge = "arduino_bridge.__main__:main"
